requirements.txt
azure-cognitiveservices-vision-computervision==0.9.0
azure-common==1.1.28
azure-core==1.31.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Flask==3.0.3
idna==3.10
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
msrest==0.7.1
oauthlib==3.2.2
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.4
